Removed components, consollidated into Dashboard.
