--- components/index.js
@@ -1,2 +1,0 @@
-export { default as Navbar } from './Navbar';
-export { default as Timer } from './Timer';
