+++ screens/index.js
@@ -1,3 +1,5 @@
 export { default as Dashboard } from './Dashboard';
-export { default as Login } from './Login';
+export { default as AuthPage } from './AuthPage';
+export { default as Login } from './AuthPage';
+export { default as SignUp } from './AuthPage';
 export { default as Welcome } from './Welcome';
